requirements.txt
pyyaml
Jinja2